migrated ItemVariable to new API
